Abstract window related logic and positioning from RoutineDisplay to WindowWrapper
