+++ src/Components/RoutineDisplay/RoutineDisplay.scss
@@ -6,11 +6,6 @@
     user-select: none;
   }
 
-  position: fixed;
-  z-index: -5;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -45%) scale(0.8);
   width: 80rem;
   height: 50rem;
   background-color: $dark-200;
@@ -42,7 +37,6 @@
 
   &.visible {
     z-index: 10;
-    transform: translate(-50%, -45%) scale(1);
     opacity: 1;
 
     ~ * {
