+++ src/Components/RoutineDisplay/RoutineDisplay.tsx
@@ -2,7 +2,7 @@ import "./RoutineDisplay.scss";
 import { useRef } from "react";
 import useOutsideClick from "../../Hooks/UseOutsideClick";
 import { useNavigate } from "react-router-dom";
-import AnimatedLayout from "../../AnimatedLayout";
+import WindowWrapper from "../WindowWrapper/WindowWrapper";
 
 export default function RoutineDisplay() {
   const routineDisplayRef = useRef<HTMLDivElement>(null);
@@ -12,13 +12,13 @@ export default function RoutineDisplay() {
   useOutsideClick(routineDisplayRef, () => navigate("/me"), "left");
 
   return (
-    <AnimatedLayout>
+    <WindowWrapper>
       <div ref={routineDisplayRef} className={`routine-display visible`}>
         <div className="routine-display-header">
           <p className="routine-display-title">Push</p>
           <button className="routine-display-edit">Edit</button>
         </div>
       </div>
-    </AnimatedLayout>
+    </WindowWrapper>
   );
 }
