+++ src/Components/WindowWrapper/AnimatedLayout.tsx
@@ -1,5 +1,6 @@
 import { motion as Motion, Variants } from "framer-motion";
 import { ReactNode } from "react";
+import "./WindowWrapper.scss";
 
 type Props = {
   children: ReactNode;
@@ -8,27 +9,15 @@ type Props = {
 const variants: Variants = {
   hidden: {
     opacity: 0,
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%) scale(0.5)",
-    zIndex: 100000,
+    scale: 0.5,
   },
   enter: {
     opacity: 1,
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%) scale(1)",
-    zIndex: 100000,
+    scale: 1,
   },
   exit: {
     opacity: 0,
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%) scale(0.5)",
-    zIndex: 100000,
+    scale: 0.5,
   },
 };
 
@@ -40,7 +29,7 @@ const AnimatedLayout = ({ children }: Props): React.JSX.Element => {
       exit="exit"
       variants={variants}
       transition={{ duration: 0.3, type: "easeInOut" }}
-      className="relative"
+      className="window-wrapper"
     >
       {children}
     </Motion.div>
