+++ src/Components/WindowWrapper/WindowWrapper.scss
@@ -1,0 +1,17 @@
+.window-wrapper {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 100vw;
+  height: 100vh;
+  z-index: 100000000;
+  translate: -50% -50%;
+  
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  & ~ * {
+    filter: blur(6px);
+  }
+}
