+++ src/index.scss
@@ -19,6 +19,7 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+  transition: filter 0.4s;
 }
 
 body {
@@ -26,7 +27,7 @@ body {
 }
 
 @media screen and (max-width: 425px) {
-  :root{
+  :root {
     font-size: 8px;
   }
 }
